refactor: extract locale resolver and formatters to util/locale.ts
Move _locale() out of CameraGalleryCard as resolveLocale(hass) and
relocate the locale-aware formatters (useAmPm, formatDateTime,
formatDay, formatTimeFromMs) from data/datetime-parsing into the
new module. data/datetime-parsing now has zero dependency on Intl
or HA types — it stays pure parsing.

Specs split alongside the move: 26 tests stay in datetime-parsing.spec
(parsing + ms↔key round-trip), 16 in util/locale.spec (resolveLocale,
useAmPm, format fallbacks, locale-dependent output).
